delete room done

+++ src/server/router/api/room/handlers/deleteRoom.js
@@ -1,7 +1,18 @@
 const Room = require('../../../../models/Room.js')
 
 module.exports = function (req, res) {
-  Room.findOneAndRemove({_id: req.params.id})
-    .then(task => res.json(task))
+  const userId = req.session.user._id
+
+  Room.findOne({_id: req.params.id})
+    .populate('companie', 'user')
+    .then(data => {
+      if (data.companie.user.toString() === userId) {
+        Room.findOneAndRemove({_id: req.params.id})
+          .then(task => res.json({success: true, data: task}))
+          .catch(err => console.log(err))
+      } else {
+        res.json({success: false, msg: 'Not owner'})
+      }
+    })
     .catch(err => console.log(err))
 }
